+++ views/forms.html
@@ -40,7 +40,7 @@
     <div class="flex items-center">
       <button type="submit" hx-indicator="#dig-indicator"
         class="rounded bg-amber-600 p-2 text-stone-900 hover:bg-amber-700 focus:ring focus:ring-amber-300">Submit</button>
-      <img id="dig-indicator" class="htmx-indicator ml-2" src="/public/rings.svg">
+      <img id="dig-indicator" class="htmx-indicator ml-2" src="/rings.svg">
     </div>
     <div id="dig-response-target"></div>
   </form>
@@ -82,7 +82,7 @@
     <div class="flex items-center">
       <button type="submit" hx-indicator="#med-indicator"
         class="rounded bg-amber-600 p-2 text-stone-900 hover:bg-amber-700 focus:ring focus:ring-amber-300">Submit</button>
-      <img id="med-indicator" class="htmx-indicator ml-2" src="/public/rings.svg">
+      <img id="med-indicator" class="htmx-indicator ml-2" src="/rings.svg">
     </div>
     <div id="med-response-target"></div>
   </form>
