+++ views/index.html
@@ -9,8 +9,8 @@
 		<script src="https://unpkg.com/htmx.org/dist/ext/remove-me.js"></script>
 		<link rel="manifest" href="/manifest.json">
 		<link rel="icon" href="/favicon.ico" type="image/x-icon">
-		<link rel="apple-touch-icon" href="/public/icon-192x192.png">
-		<link rel="apple-touch-icon" sizes="512x512" href="/public/icon-512x512.png">
+		<link rel="apple-touch-icon" href="/icon-192x192.png">
+		<link rel="apple-touch-icon" sizes="512x512" href="/icon-512x512.png">
 		<script>
   			if ('serviceWorker' in navigator) {
 			  window.addEventListener('load', () => {
